fix(Content-Recorder): ensure final generated image respects bounds after rotation when captureOnlyBackgroundImageArea is true
This resolves issue #145.

+++ lib/utils/content_recorder.dart/content_recorder_controller.dart
@@ -315,18 +315,8 @@ class ContentRecorderController {
 
         canvas.drawImageRect(
           image,
-          Rect.fromLTWH(
-            cropX.toDouble(),
-            cropY.toDouble(),
-            cropWidth.toDouble(),
-            cropHeight.toDouble(),
-          ),
-          Rect.fromLTWH(
-            0,
-            0,
-            cropWidth.toDouble(),
-            cropHeight.toDouble(),
-          ),
+          Rect.fromLTWH(cropX, cropY, cropWidth, cropHeight),
+          Rect.fromLTWH(0, 0, cropWidth, cropHeight),
           Paint(),
         );
